Fix HMR
