+++ libs/gi/ui/src/components/index.ts
@@ -15,6 +15,7 @@ export * from './InfoComponent'
 export * from './LevelSelect'
 export * from './loadout'
 export * from './NoArtWarning'
+export * from './OptTargetWrapper'
 export * from './PercentBadge'
 export * from './RefinementDropdown'
 export * from './StatDisplay'
