+++ libs/gi/ui/src/components/weapon/WeaponCardNano.tsx
@@ -18,7 +18,7 @@ import BusinessCenterIcon from '@mui/icons-material/BusinessCenter'
 import { Box, CardActionArea, Chip, Typography } from '@mui/material'
 import type { ReactNode } from 'react'
 import { useCallback, useMemo } from 'react'
-import { getCalcDisplay, resolveInfo } from '../../util'
+import { GetCalcDisplay, resolveInfo } from '../../util'
 import { CharIconSide } from '../character'
 import { WeaponNameTooltip } from './WeaponNameTooltip'
 
@@ -205,7 +205,7 @@ function WeaponStat({ node }: { node: CalcResult }) {
         >
           <span>{icon}</span>
         </BootstrapTooltip>
-        <span>{getCalcDisplay(node).valueString}</span>
+        <span>{GetCalcDisplay(node).valueString}</span>
       </Typography>
     </Box>
   )
