+++ libs/gi/ui/src/context/index.ts
@@ -1,6 +1,6 @@
 export * from './DataContext'
 export * from './FormulaDataContext'
 export * from './GraphContext'
-export * from './OptTargetContext'
+export { OptTargetContext } from './OptTargetContext'
 export * from './SillyContext'
 export * from './SnowContext'
